blockchainApp_local/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pymongo==4.6.0
python-dotenv==1.0.0
python-multipart==0.0.6
bcrypt==4.1.2
PyJWT==2.8.0
web3==6.15.1
ipfshttpclient==0.8.0a2
pydantic==2.5.2
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
email-validator==2.1.0
cryptography>=41.0.0
requests>=2.31.0